Fix #104
Deprecated function vpts() now uses the public function bind_into_vpts(), which uses the vplist_to_vpts in the background, so the result is the same

Deprecated function is.vplist() now uses sapply(c(vp, vp), is.vp) in the background, so it actually provides functionality to the user. Doc is updated.